@@ -3,15 +3,14 @@
   <Import Project="..\..\common.props" />
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>Team.Bloging</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Volo.Abp.TestBase" Version="0.18.1" />
-    <PackageReference Include="Volo.Abp.Autofac" Version="0.18.1" />
-    <PackageReference Include="Volo.Abp.Authorization" Version="0.18.1" />
-    <ProjectReference Include="..\..\src\Team.Bloging.Domain\Team.Bloging.Domain.csproj" />
+    <PackageReference Include="Volo.Abp.TestBase" Version="2.2.1" />
+    <PackageReference Include="Volo.Abp.Autofac" Version="2.2.1" />
+    <PackageReference Include="Volo.Abp.Authorization" Version="2.2.1" />
   </ItemGroup>
 
   <ItemGroup>
@@ -21,6 +20,10 @@
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.extensibility.execution" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Team.Bloging.Domain\Team.Blogging.Domain.csproj" />
   </ItemGroup>
 
 </Project>
