@@ -1,6 +1,6 @@
 ﻿using Volo.Abp.Localization;
 
-namespace Team.Bloging.Localization
+namespace Team.Blogging.Localization
 {
     [LocalizationResourceName("Bloging")]
     public class BlogingResource
