@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Volo.Abp.DependencyInjection;
 
-namespace Team.Bloging.Data
+namespace Team.Blogging.Data
 {
     /* This is used if database provider does't define
      * IBlogingDbSchemaMigrator implementation.
