@@ -1,11 +1,12 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.FeatureManagement;
 using Volo.Abp.Identity;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement;
 using Volo.Abp.TenantManagement;
+using Volo.Blogging;
 
-namespace Team.Bloging
+namespace Team.Blogging
 {
     [DependsOn(
         typeof(BlogingApplicationContractsModule),
@@ -14,6 +15,7 @@ namespace Team.Bloging
         typeof(AbpTenantManagementHttpApiClientModule),
         typeof(AbpFeatureManagementHttpApiClientModule)
     )]
+    [DependsOn(typeof(BloggingHttpApiClientModule))]
     public class BlogingHttpApiClientModule : AbpModule
     {
         public const string RemoteServiceName = "Default";
