@@ -1,10 +1,11 @@
-﻿using Volo.Abp.FeatureManagement;
+using Volo.Abp.FeatureManagement;
 using Volo.Abp.Identity;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement.HttpApi;
 using Volo.Abp.TenantManagement;
+using Volo.Blogging;
 
-namespace Team.Bloging
+namespace Team.Blogging
 {
     [DependsOn(
         typeof(BlogingApplicationContractsModule),
@@ -13,6 +14,7 @@ namespace Team.Bloging
         typeof(AbpTenantManagementHttpApiModule),
         typeof(AbpFeatureManagementHttpApiModule)
         )]
+    [DependsOn(typeof(BloggingHttpApiModule))]
     public class BlogingHttpApiModule : AbpModule
     {
         
