@@ -1,7 +1,7 @@
 ﻿using Volo.Abp.AspNetCore.Mvc.UI.Theme.Shared.Components;
 using Volo.Abp.DependencyInjection;
 
-namespace Team.Bloging.Web
+namespace Team.Blogging.Web
 {
     [Dependency(ReplaceServices = true)]
     public class BlogingBrandingProvider : DefaultBrandingProvider
