@@ -1,13 +1,13 @@
-﻿using Team.Bloging.Localization;
+﻿using Team.Blogging.Localization;
 using Volo.Abp.AspNetCore.Mvc.UI.RazorPages;
 
-namespace Team.Bloging.Web.Pages
+namespace Team.Blogging.Web.Pages
 {
     /* Inherit your PageModel classes from this class.
      */
-    public abstract class BlogingPageModel : AbpPageModel
+    public abstract class BloggingPageModel : AbpPageModel
     {
-        protected BlogingPageModel()
+        protected BloggingPageModel()
         {
             LocalizationResourceType = typeof(BlogingResource);
         }
