@@ -1,6 +1,6 @@
 ﻿@page
-@inherits Team.Bloging.Web.Pages.BlogingPage
-@model Team.Bloging.Web.Pages.IndexModel
+@inherits Team.Blogging.Web.Pages.BloggingPage
+@model Team.Blogging.Web.Pages.IndexModel
 @section styles {
     <abp-style-bundle>
         <abp-style src="/Pages/Index.css" />
