@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Text;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
@@ -7,24 +8,24 @@ using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Logging;
 using Volo.Abp;
 
-namespace Team.Bloging.Web
+namespace Team.Blogging.Web
 {
     public class Startup
     {
-        public IServiceProvider ConfigureServices(IServiceCollection services)
-        {
-            services.AddApplication<BlogingWebModule>(options =>
-            {
-                options.UseAutofac();
-            });
+      
 
-            return services.BuildServiceProviderFromFactory();
+    
+
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddApplication<BlogingWebModule>();
         }
 
         public void Configure(IApplicationBuilder app, ILoggerFactory loggerFactory)
         {
+            Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
+
             app.InitializeApplication();
-            app.UseVirtualFiles();
         }
     }
 }
