@@ -3,14 +3,14 @@ using System.Threading.Tasks;
 using Volo.Abp.Identity;
 using Xunit;
 
-namespace Team.Bloging.Samples
+namespace Team.Blogging.Samples
 {
     /* This is just an example test class.
      * Normally, you don't test code of the modules you are using
      * (like IIdentityUserAppService here).
      * Only test your own application services.
      */
-    public class SampleAppServiceTests : BlogingApplicationTestBase
+    public class SampleAppServiceTests : BloggingApplicationTestBase
     {
         private readonly IIdentityUserAppService _userAppService;
 
