@@ -3,13 +3,14 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp;
 using Volo.Abp.Modularity;
+using Volo.Abp.Testing;
 using Volo.Abp.Uow;
 
-namespace Team.Bloging
+namespace Team.Blogging
 {
     /* All test classes are derived from this class, directly or indirectly.
      */
-    public abstract class BlogingTestBase<TStartupModule> : AbpIntegratedTest<TStartupModule> 
+    public abstract class BloggingTestBase<TStartupModule> : AbpIntegratedTest<TStartupModule> 
         where TStartupModule : IAbpModule
     {
         protected override void SetAbpApplicationCreationOptions(AbpApplicationCreationOptions options)
@@ -19,30 +20,26 @@ namespace Team.Bloging
 
         protected virtual void WithUnitOfWork(Action action)
         {
-            WithUnitOfWork(new UnitOfWorkOptions(), action);
+            WithUnitOfWork(new AbpUnitOfWorkOptions(), action);
         }
 
-        protected virtual void WithUnitOfWork(UnitOfWorkOptions options, Action action)
+        protected virtual void WithUnitOfWork(AbpUnitOfWorkOptions options, Action action)
         {
-            using (var scope = ServiceProvider.CreateScope())
-            {
-                var uowManager = scope.ServiceProvider.GetRequiredService<IUnitOfWorkManager>();
+            using var scope = ServiceProvider.CreateScope();
+            var uowManager = scope.ServiceProvider.GetRequiredService<IUnitOfWorkManager>();
 
-                using (var uow = uowManager.Begin(options))
-                {
-                    action();
+            using var uow = uowManager.Begin(options);
+            action();
 
-                    uow.Complete();
-                }
-            }
+            uow.CompleteAsync();
         }
 
         protected virtual Task WithUnitOfWorkAsync(Func<Task> func)
         {
-            return WithUnitOfWorkAsync(new UnitOfWorkOptions(), func);
+            return WithUnitOfWorkAsync(new AbpUnitOfWorkOptions(), func);
         }
 
-        protected virtual async Task WithUnitOfWorkAsync(UnitOfWorkOptions options, Func<Task> action)
+        protected virtual async Task WithUnitOfWorkAsync(AbpUnitOfWorkOptions options, Func<Task> action)
         {
             using (var scope = ServiceProvider.CreateScope())
             {
@@ -59,30 +56,26 @@ namespace Team.Bloging
 
         protected virtual TResult WithUnitOfWork<TResult>(Func<TResult> func)
         {
-            return WithUnitOfWork(new UnitOfWorkOptions(), func);
+            return WithUnitOfWork(new AbpUnitOfWorkOptions(), func);
         }
 
-        protected virtual TResult WithUnitOfWork<TResult>(UnitOfWorkOptions options, Func<TResult> func)
+        protected virtual TResult WithUnitOfWork<TResult>(AbpUnitOfWorkOptions options, Func<TResult> func)
         {
-            using (var scope = ServiceProvider.CreateScope())
-            {
-                var uowManager = scope.ServiceProvider.GetRequiredService<IUnitOfWorkManager>();
+            using var scope = ServiceProvider.CreateScope();
+            var uowManager = scope.ServiceProvider.GetRequiredService<IUnitOfWorkManager>();
 
-                using (var uow = uowManager.Begin(options))
-                {
-                    var result = func();
-                    uow.Complete();
-                    return result;
-                }
-            }
+            using var uow = uowManager.Begin(options);
+            var result = func();
+            uow.CompleteAsync();
+            return result;
         }
 
         protected virtual Task<TResult> WithUnitOfWorkAsync<TResult>(Func<Task<TResult>> func)
         {
-            return WithUnitOfWorkAsync(new UnitOfWorkOptions(), func);
+            return WithUnitOfWorkAsync(new AbpUnitOfWorkOptions(), func);
         }
 
-        protected virtual async Task<TResult> WithUnitOfWorkAsync<TResult>(UnitOfWorkOptions options, Func<Task<TResult>> func)
+        protected virtual async Task<TResult> WithUnitOfWorkAsync<TResult>(AbpUnitOfWorkOptions options, Func<Task<TResult>> func)
         {
             using (var scope = ServiceProvider.CreateScope())
             {
