@@ -4,7 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Volo.Abp;
 
-namespace Team.Bloging
+namespace Team.Blogging
 {
     public class BlogingWebTestStartup
     {
